requirements.txt
fastapi==0.115.14
pandas==2.3.0
protobuf==6.31.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
